Added start of UI

+++ src/Model/CharacterStats.java
@@ -35,7 +35,7 @@ public class CharacterStats extends Stats {
      * @param difficulty the {@link Difficulty} in which the resist should be calculated
      * @return the total resistance of the specified type at the chosen difficulty
      */
-    public double getTotalResist(ResistType type, Difficulty difficulty) {
+    public int getTotalResist(ResistType type, Difficulty difficulty) {
         return getAdditionalResistValue(type) + difficulty.getResistancePenalty();
     }
 
@@ -45,7 +45,7 @@ public class CharacterStats extends Stats {
      * @param difficulty the {@link Difficulty} in which the resist should be calculated
      * @return the effective resistance of the specified type at the chosen difficulty
      */
-    public double getEffectiveResist(ResistType type, Difficulty difficulty) {
+    public int getEffectiveResist(ResistType type, Difficulty difficulty) {
         return Math.min(getTotalResist(type, difficulty), getMaxResist(type));
     }
 
@@ -54,7 +54,7 @@ public class CharacterStats extends Stats {
      * @param type the {@link ResistType} to be returned
      * @return the maximum resistance of the specified type
      */
-    public double getMaxResist(ResistType type) {
+    public int getMaxResist(ResistType type) {
         return getMaximumResistValue(type) + GameConstants.BASE_RESISTANCE_CAP;
     }
